pyproject.toml
[project]
name = "expert-network-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
